break controls.js into smaller components
